Backend/requirements.txt
Django>=5.2,<6.0
mysqlclient>=2.2,<3.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
